requirements.txt
dj-database-url==0.2.2
dj-static==0.0.5
django-otree==0.2.151
django-toolbelt==0.0.1
django-vanilla-views==1.0.2
django-countries==1.5
gunicorn==18.0
static==1.0.2
wsgiref==0.1.2
six
raven
psycopg2==2.5.1
